pyproject.toml
[tool.poetry]
name = "phone-address-api"
version = "0.1.0"
description = ""
authors = ["poligrafo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.14"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
redis = "^5.0.8"
pytest-asyncio = "^0.24.0"
pydantic-settings = "^2.4.0"
httpx = "^0.27.2"
pydantic = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
